feat: 룩 이동 규칙 테스트 및 구현
