+++ README.md
@@ -56,8 +56,9 @@
   - 적 방향 직선 1칸 이동 (처음 이동 시에는 2칸 이동 가능)
   - 공격 : 적 방향 좌, 우 대각선 1칸
 
-- [] 룩 (5)
-  - 모든 직선 방향으로 원하는 만큼 이동 가능
+- [x] 룩 (5)
+  - [x] 모든 직선 방향으로 원하는 만큼 이동 가능
+    - [x] ERROR : 룩 이동 패턴으로 이동할 수 없는 위치일 경우
 
 - [] 나이트 (2.5)
   - 모든 직선 방향 1칸 + 이동한 직선 방향의 좌, 우 대각선 1칸으로 이동 가능 (진행 방향이 가로막혀도 적, 아군 상관없이 뛰어넘을 수 있다.)
