+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,9 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.board.Position;
-
-import java.util.Set;
-
 public class Pawn extends Piece {
 
     public Pawn(final Color color) {
@@ -11,7 +7,7 @@ public class Pawn extends Piece {
     }
 
     @Override
-    public Set<Position> findPaths(final Position source, final Position target) {
-        return null;
+    protected void validatePattern(final int fileGap, final int rankGap) {
+
     }
 }
